providentia-flask/requirements.txt
APScheduler==3.6.0
Click==7.0
gremlinpython==3.4.1
Flask==1.0.3
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.2
pytz==2019.1
six==1.12.0
tzlocal==1.5.1
Werkzeug==0.15.4
